main.py
from jpype import startJVM, shutdownJVM, java, JClass

## TODO: THIS DOES NOT WORK YET

def execute_java_method():
    # Start the JVM
    startJVM(convertStrings=False)

    try:
        # Import the necessary Java classes
        MyJavaClass = JClass("moss.moss.MyJavaClass")

        # Create an instance of the Java class
        my_instance = MyJavaClass()

        # Call the Java method
        result = my_instance.myMethod("Hello from Python!")

        # Print the result
        print(f"Result from Java: {result}")

    except Exception as e:
        print(f"An error occurred: {e}")

    finally:
        # Shutdown the JVM
        shutdownJVM()


# Define the Java class
class MyJavaClass:
    def __init__(self):
        pass

    def myMethod(self, input_string):
        return f"This is the result from Java: {input_string}"


# Execute the function
execute_java_method()
