requirements.txt
JPype1==1.5.0
packaging==24.1
